added AllGraphData, data files renamed to uppercase
